Fix handling of cursors at the start/end of lines
If the cursor is at the start of the line and there's only whitespace
before it, we treat the first non-whitespace character as the start.
Consider this example where [] is the cursor:

    class Foo {
      [] fn foo {}
    }

Prior to this commit, pressing % would result in the cursor jumping to
the closing } of the class, rather than the } of the method.

In addition, if you select beyond a line's last character using visual
mode (so including the trailing newline), apply a similar technique to
ensure you don't jump to the start/end of the parent node.